pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bact-trait-cluster"
version = "1.0.0"
description = "Multi-algorithm consensus clustering with topological validation for binary genotype-phenotype matrices"
requires-python = ">=3.10"
license = {text = "MIT"}
readme = "README.md"
authors = [
    {name = "Maciej Kochanowski", email = "m.kochanowski@piwet.pulawy.pl"}
]
keywords = ["antimicrobial resistance", "clustering", "consensus clustering", "Streptococcus suis", "bioinformatics", "binary matrix", "persistent homology"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Intended Audience :: Science/Research",
]
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    "statsmodels>=0.14",
    "joblib>=1.3",
    "pyyaml>=6.0",
    "tqdm>=4.66",
]

[project.urls]
Homepage = "https://github.com/MK-vet/bact-trait-cluster"
Repository = "https://github.com/MK-vet/bact-trait-cluster"
"Bug Tracker" = "https://github.com/MK-vet/bact-trait-cluster/issues"

[project.optional-dependencies]
kmodes = ["kmodes>=0.12"]
shap = ["shap>=0.43", "lightgbm>=4.0"]
tda = ["ripser>=0.6"]
gui = ["marimo>=0.9", "plotly>=5.18", "matplotlib>=3.7"]
dev = ["pytest>=7", "pytest-cov>=4", "ruff>=0.4"]
all = ["bact-trait-cluster[kmodes,shap,tda,gui,dev]"]

[project.scripts]
bact-trait-cluster = "bacttraitcluster.cli:main"
bact-trait-cluster-dashboard = "bacttraitcluster.dashboard_launcher:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
